refactor: move previous react layer stuff into new dirs, delete the rest
