+++ packages/client/src/app/App.tsx
@@ -4,11 +4,11 @@ import { ViewToggle } from "./ui/ViewToggle";
 import { useView } from "./useView";
 import { Position } from "./ui/Position";
 import styled from "styled-components";
-import { JoinSocial } from "../layers/react/components/JoinSocial";
+import { JoinSocial } from "./ui/JoinSocial";
 import { MapLayerToggle } from "./ui/MapLayerToggle";
 import { NoaLayer } from "./NoaLayer";
 import { PhaserLayer } from "./PhaserLayer";
-import { Container } from "../layers/react/components/common";
+import { Container } from "./ui/Container";
 import { useStore } from "../store";
 import { Layer } from "./ui/Layer";
 import { useNetworkLayer } from "./useNetworkLayer";
