+++ packages/client/src/app/ui/Text.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const Gold = styled.span`
+  color: #ff0;
+`;
+export const Red = styled.span`
+  color: #eb453b;
+`;
+export const Title = styled(Gold)`
+  text-transform: capitalize;
+`;
