+++ packages/client/src/app/ui/ViewToggle.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import styled from "styled-components";
 import { TILE_HEIGHT, TILE_WIDTH } from "../../layers/phaser/constants";
-import { Container } from "../../layers/react/components/common";
+import { Container } from "./Container";
 import { useStore } from "../../store";
 import { useView } from "../useView";
