+++ packages/client/src/app/ui/icons/CheckIcon.tsx
