+++ packages/client/src/app/ui/icons/CloseIcon.tsx
