--- packages/client/src/layers/react/components/common/Border.tsx
@@ -1,5 +1,0 @@
-import styled from "styled-components";
-
-export const Border = styled.div<{ borderColor: string }>`
-  border: 3px ${(p) => p.borderColor} solid;
-`;
