--- packages/client/src/layers/react/components/index.ts
@@ -1,24 +1,0 @@
-// import { registerComponentBrowser } from "./ComponentBrowser";
-// import { registerActionQueue } from "./ActionQueue";
-// import { registerCrosshairs } from "./Crosshairs";
-import { registerLoadingState } from "./LoadingState";
-// import { registerBlockExplorer } from "./BlockExplorer";
-// import { registerInventory } from "./Inventory";
-// import { registerSidebar } from "./Sidebar";
-// import { registerPlugins } from "./Plugins";
-import { registerMapUI } from "./MapUI";
-import { registerPosition } from "./Position";
-// export * from "./common";
-
-export function registerUIComponents() {
-  registerLoadingState();
-  // registerActionQueue();
-  // registerCrosshairs();
-  // registerBlockExplorer();
-  // registerInventory();
-  // registerSidebar();
-  // registerComponentBrowser();
-  // registerPlugins();
-  registerMapUI();
-  registerPosition();
-}
