--- packages/client/src/layers/react/engine/components/MainWindow.tsx
@@ -1,7 +1,0 @@
-import React from "react";
-import { observer } from "mobx-react-lite";
-import { ComponentRenderer } from "./ComponentRenderer";
-
-export const MainWindow: React.FC = observer(() => {
-  return <ComponentRenderer />;
-});
