--- packages/client/src/layers/react/engine/components/index.ts
@@ -1,4 +1,0 @@
-export { Cell } from "./Cell";
-export { ComponentRenderer } from "./ComponentRenderer";
-export { MainWindow } from "./MainWindow";
-export { BootScreen } from "./BootScreen";
