--- packages/client/src/layers/react/engine/constants.ts
@@ -1,1 +1,0 @@
-export const ReloadEvent = new Event("reload:react");
