--- packages/client/src/layers/react/engine/hooks/useEngineStore.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { EngineContext } from "../context";
-
-export function useEngineStore() {
-  return useContext(EngineContext);
-}
