--- packages/client/src/layers/react/engine/hooks/useLayers.ts
@@ -1,6 +1,0 @@
-import { useContext } from "react";
-import { LayerContext } from "../context";
-
-export function useLayers() {
-  return useContext(LayerContext);
-}
