--- packages/client/src/layers/react/engine/index.ts
@@ -1,5 +1,0 @@
-export { Engine } from "./Engine";
-export { registerUIComponent } from "./store";
-export { useLayers } from "./hooks";
-export { BootScreen } from "./components";
-export { ReloadEvent } from "./constants";
