--- packages/client/src/layers/react/engine/store.ts
@@ -1,19 +1,0 @@
-import { observable, action } from "mobx";
-import { Observable } from "rxjs";
-import { Layers } from "../../../types";
-import { GridConfiguration, UIComponent } from "./types";
-
-export const EngineStore = observable({
-  UIComponents: new Map<string, UIComponent>(),
-});
-
-export const registerUIComponent = action(
-  <T>(
-    id: string,
-    gridConfig: GridConfiguration,
-    requirement: (layers: Layers) => Observable<T>,
-    Render: React.FC<NonNullable<T>>
-  ) => {
-    EngineStore.UIComponents.set(id, { requirement, Render: Render as React.FC, gridConfig });
-  }
-);
